Refactor Topic taxonomy block before React Query 5 update
