+++ src/containers/ArticlePage/TopicArticlePage/components/TopicArticleTaxonomy.tsx
@@ -6,36 +6,17 @@
  *
  */
 
-import { MouseEvent, useCallback, useMemo, useState } from 'react';
+import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
-import styled from '@emotion/styled';
-import partition from 'lodash/partition';
-import { ErrorMessage } from '@ndla/ui';
-import { ButtonV2 } from '@ndla/button';
-import { spacing, colors } from '@ndla/core';
-import { IUpdatedArticle, IArticle } from '@ndla/types-backend/draft-api';
-import { SingleValue } from '@ndla/select';
-import { useQueryClient } from '@tanstack/react-query';
-import { Node } from '@ndla/types-taxonomy';
-import isEqual from 'lodash/isEqual';
 import sortBy from 'lodash/sortBy';
-import differenceBy from 'lodash/differenceBy';
-import { groupChildNodes } from '../../../../util/taxonomyHelpers';
-import handleError from '../../../../util/handleError';
-import SaveButton from '../../../../components/SaveButton';
-import TopicArticleConnections from './TopicArticleConnections';
-
-import { FormikFieldHelp } from '../../../../components/FormikField';
-import TaxonomyConnectionErrors from '../../components/TaxonomyConnectionErrors';
-import { TAXONOMY_ADMIN_SCOPE } from '../../../../constants';
-import { useSession } from '../../../Session/SessionProvider';
+import { ErrorMessage } from '@ndla/ui';
+import { Spinner } from '@ndla/icons';
+import { IUpdatedArticle, IArticle } from '@ndla/types-backend/draft-api';
+import { Node } from '@ndla/types-taxonomy';
 import { useTaxonomyVersion } from '../../../StructureVersion/TaxonomyVersionProvider';
-import VersionSelect from '../../components/VersionSelect';
 import { useVersions } from '../../../../modules/taxonomy/versions/versionQueries';
-import { nodeQueryKeys, useNodes } from '../../../../modules/nodes/nodeQueries';
-import { fetchChildNodes } from '../../../../modules/nodes/nodeApi';
-import { NodeWithChildren } from '../../../../components/Taxonomy/TaxonomyBlockNode';
-import { useCreateTopicNodeConnectionsMutation } from '../../../../modules/taxonomy/taxonomyMutations';
+import { useNodes } from '../../../../modules/nodes/nodeQueries';
+import TopicTaxonomyBlock from './TopicTaxonomyBlock';
 
 type Props = {
   article: IArticle;
@@ -43,30 +24,6 @@ type Props = {
   articleLanguage: string;
   hasTaxEntries: boolean;
 };
-
-const ButtonContainer = styled.div`
-  display: flex;
-  justify-content: flex-end;
-  gap: ${spacing.xsmall};
-  margin-bottom: ${spacing.small};
-`;
-
-const InvalidPlacementsWrapper = styled.ul`
-  display: flex;
-  flex-direction: column;
-  gap: ${spacing.xsmall};
-  width: 100%;
-  padding: 0px;
-  margin: 0px 0px ${spacing.normal} ${spacing.normal};
-`;
-
-const InvalidPlacement = styled.li`
-  width: 100%;
-  color: ${colors.support.red};
-  margin: 0px;
-  margin: 0px;
-  padding: 0px;
-`;
 
 const partitionByValidity = (nodes: Node[]) => {
   const [validPlacements, invalidPlacements] = nodes.reduce<[Node[], Node[]]>(
@@ -85,17 +42,11 @@ const partitionByValidity = (nodes: Node[]) => {
 };
 
 const TopicArticleTaxonomy = ({ article, updateNotes, articleLanguage, hasTaxEntries }: Props) => {
-  const [status, setStatus] = useState('loading');
-  const [showWarning, setShowWarning] = useState(false);
-  const [isSaving, setIsSaving] = useState(false);
-  const [subjects, setSubjects] = useState<NodeWithChildren[]>([]);
   const { t } = useTranslation();
-  const { userPermissions } = useSession();
   const { taxonomyVersion, changeVersion } = useTaxonomyVersion();
-  const { data: versions } = useVersions();
-  const qc = useQueryClient();
+  const versionsQuery = useVersions();
 
-  useNodes(
+  const subjectsQuery = useNodes(
     { language: articleLanguage, taxonomyVersion, nodeType: 'SUBJECT' },
     {
       select: (subject) =>
@@ -103,143 +54,21 @@ const TopicArticleTaxonomy = ({ article, updateNotes, articleLanguage, hasTaxEnt
           subject.filter((s) => !!s.name),
           (s) => s.name,
         ),
-      onSuccess: (data) => setSubjects(data),
     },
   );
 
-  const [placements, setPlacements] = useState<Node[]>([]);
-  const [invalidPlacements, setInvalidPlacements] = useState<Node[]>([]);
+  const nodesQuery = useNodes({
+    language: articleLanguage,
+    contentURI: `urn:article:${article.id}`,
+    taxonomyVersion,
+    includeContexts: true,
+  });
 
-  const createTopicNodeConnectionsMutation = useCreateTopicNodeConnectionsMutation();
+  const [validPlacements, invalidPlacements] = useMemo(() => {
+    return partitionByValidity(nodesQuery.data ?? []);
+  }, [nodesQuery.data]);
 
-  const nodesQuery = useNodes(
-    {
-      language: articleLanguage,
-      contentURI: `urn:article:${article.id}`,
-      taxonomyVersion,
-      includeContexts: true,
-    },
-    {
-      onSuccess: (data) => {
-        const [validPlacements, invalidPlacements] = partitionByValidity(data);
-        setPlacements(validPlacements);
-        setInvalidPlacements(invalidPlacements);
-      },
-    },
-  );
-
-  const initialPlacements: Node[] = useMemo(
-    () => JSON.parse(JSON.stringify(partitionByValidity(nodesQuery.data ?? [])[0])),
-    [nodesQuery.data],
-  );
-
-  const [resources, topics] = useMemo(
-    () => partition(nodesQuery.data, (node) => node.nodeType === 'RESOURCE'),
-    [nodesQuery.data],
-  );
-
-  const isDirty = useMemo(
-    () => !isEqual(initialPlacements, placements),
-    [initialPlacements, placements],
-  );
-
-  const getSubjectTopics = useCallback(
-    async (subjectId: string) => {
-      if (subjects.some((subject) => subject.id === subjectId && !!subject.childNodes)) {
-        return;
-      }
-      try {
-        const nodes = await fetchChildNodes({
-          id: subjectId,
-          language: articleLanguage,
-          taxonomyVersion,
-          nodeType: ['TOPIC'],
-          recursive: true,
-        });
-        const childNodes = groupChildNodes(nodes);
-        setSubjects((subjects) =>
-          subjects.map((s) => (s.id === subjectId ? { ...s, childNodes } : s)),
-        );
-      } catch (err) {
-        handleError(err);
-      }
-    },
-    [articleLanguage, subjects, taxonomyVersion],
-  );
-
-  const addConnection = useCallback((node: Node) => {
-    setPlacements((placements) => placements.concat(node));
-  }, []);
-
-  const handleSubmit = async (evt: MouseEvent<HTMLButtonElement>) => {
-    evt.preventDefault();
-    setIsSaving(true);
-    const newNodes = differenceBy(placements, initialPlacements, (p) => p.id);
-    if (!newNodes.length) return;
-    try {
-      await createTopicNodeConnectionsMutation.mutateAsync({
-        placements: newNodes,
-        name: article.title?.title ?? '',
-        articleId: article.id,
-        taxonomyVersion,
-      });
-      await updateNotes({
-        revision: article.revision,
-        language: article.title?.language,
-        notes: ['Oppdatert taksonomi.'],
-      });
-      await qc.invalidateQueries({
-        queryKey: nodeQueryKeys.nodes({
-          contentURI: `urn:article:${article.id}`,
-          taxonomyVersion,
-          language: articleLanguage,
-          includeContexts: true,
-        }),
-      });
-      setIsSaving(false);
-      qc.invalidateQueries({
-        queryKey: nodeQueryKeys.nodes({
-          language: articleLanguage,
-          taxonomyVersion,
-          nodeType: 'SUBJECT',
-        }),
-      });
-    } catch (err) {
-      handleError(err);
-      setStatus('error');
-      setIsSaving(false);
-    }
-  };
-
-  const onReset = useCallback(() => {
-    if (!isDirty) {
-      return;
-    } else if (!showWarning) {
-      setShowWarning(true);
-    } else {
-      setPlacements(initialPlacements);
-      setShowWarning(false);
-      changeVersion('draft');
-    }
-  }, [changeVersion, initialPlacements, isDirty, showWarning]);
-
-  const onVersionChanged = useCallback(
-    (newVersion: SingleValue) => {
-      if (!newVersion || newVersion.value === taxonomyVersion) return;
-      const oldVersion = taxonomyVersion;
-      changeVersion(newVersion.value);
-      setShowWarning(false);
-      qc.removeQueries({
-        predicate: (query) => {
-          const qk = query.queryKey as [string, Record<string, any>];
-          return qk[1]?.taxonomyVersion === oldVersion;
-        },
-      });
-    },
-    [changeVersion, qc, taxonomyVersion],
-  );
-
-  if (status === 'error') {
+  if (nodesQuery.isError || subjectsQuery.isError || versionsQuery.isError) {
     changeVersion('');
     return (
       <ErrorMessage
@@ -255,61 +84,27 @@ const TopicArticleTaxonomy = ({ article, updateNotes, articleLanguage, hasTaxEnt
         }}
       />
     );
+  } else if (
+    nodesQuery.isInitialLoading ||
+    subjectsQuery.isInitialLoading ||
+    versionsQuery.isInitialLoading
+  ) {
+    return <Spinner />;
   }
 
-  const isTaxonomyAdmin = userPermissions?.includes(TAXONOMY_ADMIN_SCOPE);
-
   return (
-    <>
-      {!hasTaxEntries && <FormikFieldHelp error>{t('errorMessage.missingTax')}</FormikFieldHelp>}
-      {isTaxonomyAdmin && (
-        <>
-          <TaxonomyConnectionErrors
-            articleType={article.articleType ?? 'topic-article'}
-            resources={resources}
-            topics={topics}
-          />
-          <VersionSelect versions={versions ?? []} onVersionChanged={onVersionChanged} />
-        </>
-      )}
-      <TopicArticleConnections
-        addConnection={addConnection}
-        structure={subjects}
-        selectedNodes={placements}
-        getSubjectTopics={getSubjectTopics}
-      />
-      {!!invalidPlacements.length && isTaxonomyAdmin && (
-        <details>
-          <summary>{t('errorMessage.invalidTopicPlacements')}</summary>
-          <InvalidPlacementsWrapper>
-            {invalidPlacements.map((placement) => (
-              <InvalidPlacement key={placement.id}>{placement.id}</InvalidPlacement>
-            ))}
-          </InvalidPlacementsWrapper>
-        </details>
-      )}
-      {showWarning && <FormikFieldHelp error>{t('errorMessage.unsavedTaxonomy')}</FormikFieldHelp>}
-      <ButtonContainer>
-        <ButtonV2
-          variant="outline"
-          onClick={onReset}
-          disabled={!isDirty || createTopicNodeConnectionsMutation.isLoading}
-        >
-          {t('reset')}
-        </ButtonV2>
-        <SaveButton
-          isSaving={isSaving}
-          showSaved={createTopicNodeConnectionsMutation.isSuccess && !isDirty}
-          disabled={!isDirty || nodesQuery.isFetching}
-          onClick={handleSubmit}
-          defaultText="saveTax"
-          formIsDirty={isDirty}
-        />
-      </ButtonContainer>
-      <FormikFieldHelp float="right" warning>
-        {t('warningMessage.taxonomy')}
-      </FormikFieldHelp>
-    </>
+    <TopicTaxonomyBlock
+      key={taxonomyVersion}
+      article={article}
+      subjects={subjectsQuery.data ?? []}
+      nodes={nodesQuery.data ?? []}
+      validPlacements={validPlacements}
+      invalidPlacements={invalidPlacements}
+      versions={versionsQuery.data ?? []}
+      hasTaxEntries={hasTaxEntries}
+      articleLanguage={articleLanguage}
+      updateNotes={updateNotes}
+    />
   );
 };
